# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SynTrends Agent API",
    "description": "Agent API + human onboarding web + persistence + KYC provider (Phase F).",
    "version": "0.2.0"
  },
  "paths": {
...
    "/testnet/faucet": {
      "get": {
        "tags": [
          "testnet"
        ],
        "summary": "Faucet Info",
        "operationId": "faucet_info_testnet_faucet_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "testnet"
        ],
        "summary": "Claim Faucet",
        "description": "Credit simulated testnet fiat to the agent bound to your API key.\n\nRate-limited per agent_id by ``FAUCET_COOLDOWN_SECONDS`` (default 1 hour).\nRequires an agent key \u2014 get one from the owner portal after KYC.",
        "operationId": "claim_faucet_testnet_faucet_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}